futility: updater: Allow empty WL tag in VPD when write protection is off
When WL tag is missing, we had a logic that if write protection is
enabled, which implies device have been already shipped, then we will
simply use whatever key is set. This allows shipped device to keep
receiving AU even if they didn't have right keys set. And it implies a
check that WL units in factory when entering the last "re-key"
update_firmware test, the updater will make sure WL tag is set.

However, for dogfood and lab machines, they need to update firmware
without WL tag set, and with write protection turned off. This conflicts
with the logic above.

A new logic is to allow empty WL when write protection is off, because:
 - Since we have QUIRK_ALLOW_EMPTY_WLTAG now, devices shipped with wrong
   key should be addressed, reported and enable that quirk explicitly.
 - It is true the old logic may help final update_firmware to ensure WL
   tags, but it is also possible that partner may forget to include such
   test in test list, or have it run with 'factory' mode (or 'rw-only').
   In fact, the update_firmware defaults to --force and --mode=factory
   (if RO update is allowed, which is also the case for re-key to WL)
   so the old logic didn't help. It seems more reasonable to defer the
   responsibility to finalization.

BUG=b:126800200
TEST=make futil; tests/futility/run_test_scripts.sh $(pwd)/build/futility
BRANCH=none

Change-Id: I7249e3fb1a2b7ab8ed281d2aa317aee6cde8f8db
Signed-off-by: Hung-Te Lin <hungte@chromium.org>