Fix memory leaks in cgpt
DriveOpen mallocs a couple of buffers, but DriveClose only freed them if a
particular flag was passed causing GptSave to be called. Move the free calls
out of GptSave so that the buffers are always freed, and add DriveClose calls
to a couple of cgpt functions that are missing them.

BUG=chromium:940543
TEST=precq passes, manual testing with cgpt/valgrind shows that memory leaks
     for cgpt find, show are fixed.
BRANCH=none

Change-Id: I58aeddfa6b8b4715ba4f8e064e95a660371a01c9