Update patch set 1
Patch Set 1:

I don't think there's a reason to ever delete boards out of this list? You're right that the board is dead, but we did create pre-release keys for it at some point, so if anyone ever stumbles upon an old mystery image signed with those wouldn't it be better if this tool can recognize it?

Patch-set: 1
CC: Gerrit User 1002133 <1002133@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>