vboot: refactor workbuf alloc and free code
Move workbuf allocation and freeing to separate internal
vboot functions.

These functions are called in vb2_kernel_setup and
vb2_kernel_cleanup.

This is in preparation for them to be called higher up the
stack from the vboot library caller, instead of by vboot
internally.  They may be exported as part of vboot's API,
or relocated to the caller (depthcharge) -- TBA.

BUG=b:124141368
TEST=make clean && make runtests
BRANCH=none

Change-Id: Ibaa70f62c660d46cc083a5e55a73b961eb813649
Signed-off-by: Joel Kitching <kitching@google.com>