vboot: de-externalize vb2_digest_buffer function
Clients which wish to use this function may simply import the
vb2_sha.h header to make all SHA library functionality available.

Whether or not to make a full vb2api_ set of functions for SHA
library should be considered in the future, but individual
functions should not be added to the API as was done.

BUG=b:124141368, chromium:956474
TEST=make clean && make runtests
BRANCH=none

Change-Id: If37bf753cd8cfa8122f64d864f5f23e0d969b7ce
Signed-off-by: Duncan Laurie <dlaurie@google.com>
Original-Commit-Id: de30df731edc7ed2729982de588e0f8b9082cf3c
Original-Change-Id: I1dc8ed84ecfb621a57411975f312e96a695b68f4
Original-Signed-off-by: Joel Kitching <kitching@google.com>
Original-Cq-Depend: chromium:1583942
Original-Reviewed-on: https://chromium-review.googlesource.com/1583822
Original-Commit-Ready: Joel Kitching <kitching@chromium.org>
Original-Tested-by: Joel Kitching <kitching@chromium.org>
Original-Reviewed-by: Julius Werner <jwerner@chromium.org>
Original-Reviewed-by: Simon Glass <sjg@chromium.org>