vboot: de-externalize vb2_digest_buffer function
Clients which wish to use this function may simply import the
vb2_sha.h header to make all SHA library functionality available.

Whether or not to make a full vb2api_ set of functions for SHA
library should be considered in the future, but individual
functions should not be added to the API as was done.

BUG=b:124141368, chromium:956474
TEST=make clean && make runtests
BRANCH=none

Change-Id: I1dc8ed84ecfb621a57411975f312e96a695b68f4
Signed-off-by: Joel Kitching <kitching@google.com>
Cq-Depend: chromium:1583942