vboot: add `used` property to vb2_workbuf
When using the global workbuf, we may temporarily point to it
using a local vb2_workbuf instance.  If we would like to keep the
items stored on the global workbuf permanently, we must first
know how much memory is being used, and then increment
vb2_context.workbuf_used.

Add the `used` property to vb2_workbuf to give us this
information.

Motivation: Saving GBB data on the global workbuf.

BUG=b:124141368, chromium:954774
TEST=make clean && make runtests
BRANCH=none

Change-Id: Id2971a32880ab6d2925301fc2f5cf154eab0fb98
Signed-off-by: Joel Kitching <kitching@google.com>