vboot: add vb2_workbuf_to_ctx function
An acceptable workflow for using memory from the context workbuf
is to make a local vb2_workbuf instance which points to the
context workbuf memory.  Allocations done to the local copy don't
change the passed-in work buffer, and will effectively be freed
when the local copy goes out of scope.

In some cases, we wish to retain this piece of memory in the
context workbuf.  In this case, the vb2_workbuf_to_ctx function
can be used to extend the context workbuf to encompass the memory
in use by the local workbuf.

BUG=b:124141368, chromium:954774
TEST=make clean && make runtests
BRANCH=none

Change-Id: I0c5867fa566f54870dd1363cce8bd091aee4ff85
Signed-off-by: Joel Kitching <kitching@google.com>