Update patch set 8
Patch Set 8:

Is this saving the GCC info from the firmware verification stage so that it can be used in kernel verification? If so, it doesn't seem worth it to me. It's tiny and we can read it again.

Patch-set: 8
Reviewer: Gerrit User 1165921 <1165921@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>