Update patch set 3
Patch Set 3:

> Patch Set 3: Code-Review+2
> 
> Sorry, I don't expect people to set Commit-Queue before Code-Review, so when I see someone set Commit-Queue+2 in an email I just ignore it and assume it's already good to go. (I wish it would tell you what flags you currently have set in the email already...)

Is this a roundabout way of telling me to stop doing that? =)
It's my way of dealing with a sluggish CQ and timezone differences... but if it is a really bad idea I can stop doing it.  Let me know.

Patch-set: 3
Reviewer: Gerrit User 1126709 <1126709@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>