vboot: bump VB2_SHARED_DATA_VERSION_MINOR
coreboot compiles with CL:1521406 are failing on the check:
  sd->struct_version_minor < VB2_SHARED_DATA_VERSION_MINOR

This is because coreboot uses -Werror=type-limits, and
as an unsigned int, struct_version_minor may never be <0.

This could be dealt with by removing the -Werror flag, or by
adding a compile-time conditional.  But we may as well avoid this
extra bookkeeping by simply bumping the minor version to 1.

BUG=chromium:960270
TEST=Attempt compiling coreboot with this patch applied
BRANCH=none

Change-Id: Id302372ab5fefb0728ea402b34db13fb9bb44e5c
Signed-off-by: Joel Kitching <kitching@google.com>