Update patch set 1
Patch Set 1: Code-Review-1

Sorry, I only saw this after already uploading CL:1598720

We shouldn't modify code to accommodate the warning (at least not in a meaningful way), that was not the intent of turning it on. It's useful in general but there can always be cases where it's a false positive and we really want to have the zero-check in there (there were some in coreboot too). In some cases you can fix it with a cast, if not let's just disable it with a pragma.

Patch-set: 1
Reviewer: Gerrit User 1002133 <1002133@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=-1