futility: Unify logging macros
The firmware updater introduced its own logging macros, but the rest of
futility already used a (smaller) set of macros previously. Unify the
two so that all parts of the binary use the same system.

Note that the same debug output infrastructure can (and already could
before this patch, although it was less obvious) be enabled with both
futility --debug update and futility update -d. This is a bit weird but
shouldn't really hurt and I presume we may want it for backwards
compatibility in the updater.

BRANCH=None
BUG=None
TEST=make runtests

Change-Id: I02d53c2036b533f9e8f83bca0f2837065bcc2361
Signed-off-by: Duncan Laurie <dlaurie@google.com>
Original-Commit-Id: 88a47ff99952bb4f270a4e80c80c578e39fb9477
Original-Change-Id: I8b7c1677bcef9e9772ee666c72958d27139b36a2
Original-Signed-off-by: Julius Werner <jwerner@chromium.org>
Original-Reviewed-on: https://chromium-review.googlesource.com/1601678
Original-Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Original-Reviewed-by: Hung-Te Lin <hungte@chromium.org>