Update patch set 2
Patch Set 2:

(1 comment)

> If the goal here is to "unify", then what is the reason for not relocating the other macros (ERROR, WARN, INFO, STATUS, ASPRINTF) to 2common.h?
> 
> Assuming that we do want to keep them separate, it seems quite bizarre that the ones defined locally (ERROR, WARN, etc.) use fprintf, whereas the VB2_DEBUG macro uses vb2ex_printf.

Okay, so we have several different contexts for vboot code: firmware code, generic host utility/library code (e.g. crossystem/cgpt), futility code, and code that's shared between all of them. For everything other than futility code, VB2_DEBUG() is the only logging facility atm. It's implemented via vb2ex_printf().

futility used to have its own logging facilities with Error() and Debug(). However, futility also calls into code that's shared with firmware, and the VB2_DEBUG() calls from there actually used to go nowhere at all (because they used to be guarded by a macro that wasn't set for futility). So I changed that in CL:1504140 because I thought it's confusing and not being able to see the VB2_DEBUG() output without recompiling was a problem. Then we only had the futility-only Error() and the generally available VB2_DEBUG().

Now Hung-Te added all these other logging facilities for the updater, and I get the desire for more granular error levels, but I don't think every single futility command should have its own logging macros. So I'm trying to clean some of that up with this patch, so that the ERROR/WARN/INFO/STATUS are available throughout all of futility. VB2_DEBUG() still remains the only thing that's really available in both futility and other vboot code, so that's why that's backed by vb2ex_printf() and the others suffice with fprintf. (ASPRINTF() is sort of a separate thing -- it's main purpose isn't logging -- so I didn't touch it here. I wouldn't object to making it more widely available but I think that's a separate issue.)

If we want to pull all of those other logging facilities into the rest of vboot as well, I guess we could do that, but that's a bigger deal. We'd probably want to somehow shoehorn the error level into the vb2ex_printf() API, or change the whole thing so that the prefix string is added by vboot and not handled separately in the calling firmware. All things we could discuss and do, I just didn't want to go that far with this patch yet... I was just trying to cleanup the discrepancy between different futility commands, because some part of it was blocking my next patch (I think it was the double declaration of extern int debugging_enabled;).

Patch-set: 2
Reviewer: Gerrit User 1002133 <1002133@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>