Update patch set 2
Patch Set 2:

Not sure if we really want this? crossystem isn't just for system code that needs to access these, it's also the debugging tool for engineers. I think almost every other NVRAM field can be accessed through it, it would be weird if this one is the only one that's hidden. I'd consider just renaming it.

Patch-set: 2
CC: Gerrit User 1002133 <1002133@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>