vboot: Make sure the .developer_mode file is present is switch from dev mode is displayed
The file ".developer_mode" is created by clobber only when switch to
developer mode is done and the device is reset.

There are cases when the switch to dev mode is not finished though, when
it is disabled or when the user decides to cancel the switch.

In that case the screen that the device will switch from dev mode and
the data will be cleared is displayed. To be consistent with what the
screen is displaying, we create the .developer_mode file here so that at
startup the device will indeed clear all the data.

BUG=chromium:959160
BRANCH=None
TEST=Manual

Change-Id: Ia669196addd3b793128f633f12bb16e3f6d01854