Update patch set 3
Patch Set 3:

Thank you for input Julius, I was planning to discuss with owners if such an approach would work.

My plan is to create the .developer_mode file when the screen that notifies the user about the  switch from devmode to verified boot is displayed. This way it, we would be sure that the device will indeed clobber after reboot, as the message on the screen says.

I didn't test this, but since you say there's no file system at this point, it will not work. I will have to look for other ways to make this happen.

Patch-set: 3
Reviewer: Gerrit User 1144419 <1144419@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>