Update patch set 3
Patch Set 3:

> Then searching for VB_SCREEN_TO_NORM_CONFIRMED leads me to the code from this CL. Isn't a call from here, like
> VbDisplayScreen(ctx, cparams, VB_SCREEN_TO_NORM_CONFIRMED, 0);
> displaying the screen with text from VB_SCREEN_TO_NORM_CONFIRMED? If that's the case, would saving a flag in crossystem, instead of creating the .developer_mode file, be acceptable? If it would work, I could then check the crossystem flag in chromeos_startup to decide on clobbering or not.

It is a multi-step process. The data isn't actually cleared at the time it says "local data will be cleared". That just sets the clear_tpm_owner_request NVRAM flag and reboots the system. Then, in early firmware after the reboot, the TPM owner is cleared. This doesn't delete any data but it throws away the encryption keys for it. After the system boots to userspace, chromeos_startup will detect that the TPM owner has been cleared (the is_tpm_owned part) and deduct what happened and whether data needs to be deleted based on that.

I think you're approaching this from the wrong angle. We definitely do lose access to data when the TPM owner is cleared. If you have noticed data that can survive this when you think it shouldn't (e.g. sounds from your bug like this is about shared WiFi passwords?), the main question is why that data was not encrypted with TPM-backed keys. This is a security issue so just adding new code to explicitly delete it at some random other point in the flow may not be the right fix. We need to make sure that access to all data that should be clobbered is lost the moment we clear the TPM owner, even if actual deletion of the physical bytes happens later.

Do you know where this data that you see surviving is stored? If it's under /mnt/stateful_partition/encrypted it shouldn't be possible for it to survive. If it's stored somewhere else, we need to figure out why it is there and whether it should maybe be stored in the encrypted part instead.

Patch-set: 3