Update patch set 7
Patch Set 7:

go/vboot2-oprom-cleanup doesn't seem to exist.

DISPLAY_AVAILABLE does not seem to match the previous behavior.  OPROM_MATTERS indicated that the platform cared about the oprom, and in some cases this was used to set the flag to init the display for the next reboot.  Now, it ends up checking whether the display was initialized this boot (and is therefore available) which is a very different thing...

Patch-set: 7
CC: Gerrit User 1000423 <1000423@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>