Update patch set 7
Patch Set 7:

(1 comment)

> DISPLAY_AVAILABLE contains all it needs to know, with that it doesn't need to know whether OPROM_MATTERS explicitly

Guess I'm eating my words now, but I hope we can still make the new solution work with a few tweaks. The diagnostic feature has the unusual problem that we're in recovery mode but want to request display for the next boot that's not in recovery mode, which I think doesn't happen for any other case.

Patch-set: 7