Update patch set 3
Patch Set 3:

> That device is either discovering that .developer_mode file is present, or that the TPM is not owned. Currently I've added some clobber logs to be able distinguish between the two cases, but having a redundant bit of information that the device showed the screen with switch to verified boot would help. It would allow at least to identify if there's inconsistency, and probably rely only on this bit of information in the future.

Sorry, I don't understand your problem. That the TPM was cleared (i.e. is unowned) is the bit of information you need. Adding extra redundant bits to that just opens risks for things to get out of sync.

It seems that your main problem is differentiating between OOBE (i.e. pristine or already clobbered) and transition back from dev mode, and you're checking for the presence of the install attributes file to distinguish those. Isn't that enough? Sounds like that should work and catch all cases (or are there cases where the install attributes file doesn't exist even though there is other valid data on the stateful partition?).

(BTW I noticed in CL:1075274 that you said it adds up to 80ms to the critical normal boot path. That is a pretty large amount of time compared to how aggressively we usually try to optimize firmware. Is there no way you can mitigate that... e.g. by kicking off the TPM ownership check earlier and letting it finish in the background?)

Patch-set: 3