Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > That device is either discovering that .developer_mode file is present, or that the TPM is not owned. Currently I've added some clobber logs to be able distinguish between the two cases, but having a redundant bit of information that the device showed the screen with switch to verified boot would help. It would allow at least to identify if there's inconsistency, and probably rely only on this bit of information in the future.
> 
> Sorry, I don't understand your problem. That the TPM was cleared (i.e. is unowned) is the bit of information you need. Adding extra redundant bits to that just opens risks for things to get out of sync.

I was suspecting that sometimes we get false negatives on checking if TPM is owned. Turns out it's possible that indirectly we end up in that position, when the request to TPM returns an error. Hopefully https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1626095 will fix this case, though we end up with another possible disadvantage - a risk that data on disk will not be cleared sometimes, I expect very rarely.

In short, you are right - knowing if TPM is owned is enough for me. It just seems, it's
possible to get an error on that check.

> 
> It seems that your main problem is differentiating between OOBE (i.e. pristine or already clobbered) and transition back from dev mode, and you're checking for the presence of the install attributes file to distinguish those. Isn't that enough? Sounds like that should work and catch all cases (or are there cases where the install attributes file doesn't exist even though there is other valid data on the stateful partition?).

Install attributes are created when device ownership becomes known - that is either the
device gets enrolled, or the first user login happens. So, you could take a new device and
use only guest mode on it and there will be no install attributes. At the same time only
data related to login screen basically will be in stateful partition. So, while not perfect, it's an acceptable solution to check the install attributes.

> 
> (BTW I noticed in CL:1075274 that you said it adds up to 80ms to the critical normal boot path. That is a pretty large amount of time compared to how aggressively we usually try to optimize firmware. Is there no way you can mitigate that... e.g. by kicking off the TPM ownership check earlier and letting it finish in the background?)

Good question, I would love to get rid of those 80ms on boot. I don't know it's possible to
make that check async. But, it's one more reason to see if we can create that .developer_mode file earlier so that we don't need to check the TPM at boot. I wonder if we can create it when
TPM is clearing. Or, if we can check the NVRAM instead of TPM at boot to see that a switch was done.

Patch-set: 3
Reviewer: Gerrit User 1144419 <1144419@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>