Update patch set 3
Patch Set 3:

> In short, you are right - knowing if TPM is owned is enough for me. It just seems, it's
> possible to get an error on that check.

Well, this shouldn't happen anyway (and if it does it would be important to track down the root cause), but I'd recommend to try to fix that with retries first. Commented on the other CL.

> > (BTW I noticed in CL:1075274 that you said it adds up to 80ms to the critical normal boot path. That is a pretty large amount of time compared to how aggressively we usually try to optimize firmware. Is there no way you can mitigate that... e.g. by kicking off the TPM ownership check earlier and letting it finish in the background?)
> 
> Good question, I would love to get rid of those 80ms on boot. I don't know it's possible to
> make that check async.

You could definitely kick off a background job to copy that sysfs file to somewhere in /tmp (or maybe /run) at the top of chromeos_startup and then call 'wait' when you first try to check it. Whether enough other things happen in between to cover the full 80ms I don't know, but it would be a start.

> But, it's one more reason to see if we can create that .developer_mode file earlier so that we don't need to check the TPM at boot. I wonder if we can create it when
> TPM is clearing. Or, if we can check the NVRAM instead of TPM at boot to see that a switch was done.

Firmware changes are not a good solution for this because they can not be rolled out to old devices (developer mode transition is part of read-only firmware). NVRAM isn't particularly fast anyway. We should stick with the TPM, and maybe look into ways how we can parallelize TPM operations with the rest of the boot path even further (e.g. kernel could just start prefetching some TPM data in a background thread right after the device finishes probing).

Patch-set: 3