Update patch set 4
Patch Set 4:

Sorry, what are the changes in this newest patch set for? The code reads very confusing now. Is this just to disable the display when we had EC software sync but no AUX firmware sync? Can't you just check need_wait_screen for that? Or just wait for Joel's CL to go in first which gets rid of all of that anyway.

Patch-set: 4
Reviewer: Gerrit User 1002133 <1002133@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>