vboot: correct VB2_SECDATAK_SIZE constant to 13
This constant was incorrectly set to 14, whereas
sizeof(vb2_secdatak) is 13.

Update the constant its correct value, and add a test for
each of secdata and secdatak to check the constant values.

Previously, this constant was not used anywhere.  The
secdatak space is created and initialized in coreboot
with a hard-coded (correct) size and initial value.
So there should be no worry about devices out in the
field with TPM secdatak size set to 14.

BUG=b:124141368, chromium:972956
TEST=make clean && make runtests
BRANCH=none

Change-Id: I0a6072aef2de401bd3cd40ac3b002f754da19560
Signed-off-by: Joel Kitching <kitching@google.com>