vboot/secdata: centralize secdatak reads/writes
In current kernel verification code, secdatak reads and writes
are spread throughout the code.  Eventually, we'd like to use
vb2_context.secdatak for storing the state of secdatak, and
have the caller read/save this field when necessary.

To move in that direction, centralize secdatak reads/writes into
the function vb2_flush_secdata, which should be called alongside
vb2_kernel_cleanup.

Next, functions which directly read/write to the TPM field should
be modified to simply edit the value of vb2_context.secdatak.
This will be implemented in a subsequent CL.

BUG=b:124141368, chromium:972956
TEST=make clean && make runtests
BRANCH=none

Change-Id: Ifbfb21122af0bf85e22a6d3a0d48a1db7f7c25b7
Signed-off-by: Joel Kitching <kitching@google.com>