Update patch set 2
Patch Set 2:

Why do you need this? The problem in the linked bug is exactly the one I mentioned in CL:1740664, but I think we should solve it the way I outlined there (by moving the peripheral busy check into the check_hash() callback). I don't see what it has to do with the code you're removing here (that code is only an optimization, but I think one worth keeping... if there's anything with the EC RO that prevents it from correctly determining whether we need an update, we're just skipping the optimization and need an extra reboot, but we'll check again later either way).

Patch-set: 2
CC: Gerrit User 1002133 <1002133@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>